requirements.txt
djangorestframework==3.14.0
drf-firebase-auth=1.0.0
passlib=1.7.4